admin/requirements.txt
streamlit==1.55.0
apscheduler==3.11.2
openai==2.29.0
pydantic==2.12.5
PyYAML==6.0.3
